refactor -> move LobbyInfoBox, RoomButtonBox to box package
