+++ ChatterClient/src/main/java/com/codecool/chatter/view/LobbyView.java
@@ -3,6 +3,8 @@ package com.codecool.chatter.view;
 import com.codecool.chatter.ChatterClient;
 import com.codecool.chatter.model.Lobby;
 import com.codecool.chatter.model.User;
+import com.codecool.chatter.view.box.LobbyInfoBox;
+import com.codecool.chatter.view.box.RoomButtonBox;
 import javafx.event.EventHandler;
 import javafx.geometry.Insets;
 import javafx.scene.input.MouseEvent;
