+++ ChatterClient/src/main/java/com/codecool/chatter/view/box/LobbyInfoBox.java
@@ -1,4 +1,4 @@
-package com.codecool.chatter.view;
+package com.codecool.chatter.view.box;
 
 import com.codecool.chatter.model.Lobby;
 import com.codecool.chatter.model.Room;
